requirements.txt
python-telegram-bot==20.3
pytesseract
opencv-python
Pillow
requests
httpx
apscheduler
